Refactor lenses, use better module structure
