+++ test/Chainweb/Test/Pact/SPV.hs
@@ -10,7 +10,6 @@
 {-# LANGUAGE TypeApplications #-}
 {-# LANGUAGE TypeFamilies #-}
 -- |
--- Module: Chainweb.Test.CutDB.Test
 -- Copyright: Copyright © 2018 - 2020 Kadena LLC.
 -- License: MIT
 -- Maintainer: Lars Kuhtz <lars@kadena.io>, Emily Pillmore <emily@kadena.io>
@@ -33,14 +32,11 @@ import Control.Concurrent.MVar
 import Control.Exception (SomeException, finally)
 import Control.Monad
 import Control.Lens hiding ((.=))
-import Control.Monad.Trans.Except
 
 import Data.Aeson as Aeson
 import qualified Data.ByteString.Base64.URL as B64U
-import qualified Data.ByteString.Lazy as BL
 import qualified Data.ByteString.Char8 as B8
 
-import qualified Data.Binary as Binary
 import Data.ByteString.Lazy (toStrict)
 import qualified Data.HashMap.Strict as HM
 import Data.IORef
@@ -72,7 +68,6 @@ import Pact.Types.PactValue
 import Pact.Types.Runtime (toPactId)
 import Pact.Types.SPV
 import Pact.Types.Term
-import Pact.Types.Util (AsString(..))
 
 -- internal chainweb modules
 
@@ -83,8 +78,6 @@ import Chainweb.BlockHeight
 import Chainweb.Cut
 import Chainweb.Graph
 import Chainweb.Miner.Pact
-import Chainweb.Pact.SPV
-import Chainweb.Pact.SPV.Hyperlane
 import Chainweb.Pact.Backend.Types
 import Chainweb.Payload
 import Chainweb.Payload.PayloadStore
@@ -99,6 +92,8 @@ import Chainweb.Transaction
 import Chainweb.Utils hiding (check)
 import Chainweb.Version as Chainweb
 import Chainweb.WebPactExecutionService
+
+import qualified Chainweb.Test.Pact.SPV.Hyperlane as Hyperlane
 
 import Chainweb.Storage.Table (casLookupM)
 
@@ -118,15 +113,8 @@ tests = testGroup "Chainweb.Test.Pact.SPV"
     , testCaseSteps "wrong chain execution fails" wrongChain
     , testCaseSteps "invalid proof formats fail" invalidProof
     , testCaseSteps "wrong target chain in proofs fail" wrongChainProof
-    , testGroup "hyperlane"
-      [ testCase "empty object" hyperlaneEmptyObject
 
-      , testCase "encodeTokenMessageERC20" hyperlaneEncodeTokenMessageERC20
-      , testCase "decodeTokenMessageERC20" hyperlaneDecodeTokenMessageERC20
-      , testCase "wordToDecimal" hyperlaneWordToDecimal
-
-      , testCase "encodeHyperlaneMessage" hyperlaneEncodeHyperlaneMessage
-      ]
+    , Hyperlane.tests
     ]
 
 testVer :: ChainwebVersion
@@ -149,69 +137,6 @@ _handle' e =
     let
       s = show e
     in logg System.LogLevel.Error (pack s) >> return (False, s)
-
-
-hyperlaneEmptyObject :: Assertion
-hyperlaneEmptyObject = do
-  res <- runExceptT $ evalHyperlaneCommand $ mkObject []
-  assertEqual "should fail with missing command name" (Left "Unknown hyperlane command") res
-
-hyperlaneEncodeTokenMessageERC20 :: Assertion
-hyperlaneEncodeTokenMessageERC20 = do
-  let
-    res = encodeTokenMessageERC20 $ mkObject
-        [ ("recipient", tStr $ asString ("recipient" :: Text))
-          , ("amount", tLit $ LDecimal 3333333333333333333.333333333333333333) ]
-  case res of
-    Nothing -> assertFailure "Should get the result"
-    Just t ->
-      let
-        expectedMessage :: Text = "0x0000000000000000000000000000000000000000000000000000000000000040000000000000000000000000000000000281fa059c9e179daafc1235555555550000000000000000000000000000000000000000000000000000000000000009726563697069656e740000000000000000000000000000000000000000000000"
-      in assertEqual "should get encoded message" expectedMessage t
-
-hyperlaneDecodeTokenMessageERC20 :: Assertion
-hyperlaneDecodeTokenMessageERC20 = do
-  let
-    encodedMessage :: Text = "0x0000000000000000000000000000000000000000000000000000000000000040000000000000000000000000000000000281fa059c9e179daafc1235555555550000000000000000000000000000000000000000000000000000000000000009726563697069656e740000000000000000000000000000000000000000000000"
-  encodedBinary <- case decodeHex encodedMessage of
-      Right r -> pure r
-      Left _ -> assertFailure "hyperlaneDecodeTokenMessageERC20: failed to decode"
-  let
-    tm = Binary.decode $ BL.fromStrict encodedBinary
-    expectedObject = TokenMessageERC20 "recipient" 3333333333333333333333333333333333333
-  assertEqual "Should properly decode the object" expectedObject tm
-
-hyperlaneWordToDecimal :: Assertion
-hyperlaneWordToDecimal = assertEqual "" 3333333333333333333.333333333333333333 (wordToDecimal 3333333333333333333333333333333333333)
-
-hyperlaneEncodeHyperlaneMessage :: Assertion
-hyperlaneEncodeHyperlaneMessage = do
-  let
-    obj' = mkObject
-        [ ("message", obj
-          [ ("version", tLit $ LInteger 1)
-          , ("nonce", tLit $ LInteger 1223)
-          , ("originDomain", tLit $ LInteger 626)
-          , ("sender", tStr $ asString ("0x23" :: Text))
-          , ("destinationDomain", tLit $ LInteger 8)
-          , ("recipient", tStr $ asString ("0x71c7656ec7ab88b098defb751b7401b5f6d8976f" :: Text))
-          , ("tokenMessage", obj
-              [ ("recipient", tStr $ asString ("recipient1" :: Text))
-              , ("amount", tLit $ LDecimal 3333333333333333333.333333333333333333) ]
-            )
-          ])
-        ]
-  res <- runExceptT $ evalHyperlaneCommand obj'
-  case res of
-    Left err -> assertFailure $ "Should get the result" ++ show err
-    Right o ->
-      let
-        expectedMessage :: Text = "0x01000004c70000027200000000000000000000000000000000000000000000000000000000000000230000000800000000000000000000000071c7656ec7ab88b098defb751b7401b5f6d8976f0000000000000000000000000000000000000000000000000000000000000040000000000000000000000000000000000281fa059c9e179daafc123555555555000000000000000000000000000000000000000000000000000000000000000a726563697069656e743100000000000000000000000000000000000000000000"
-        expectedObject = mkObject
-          [ ("encodedMessage", tStr $ asString expectedMessage)
-          , ("messageId", tStr $ asString ("0xd7d22abdd68430821df0732b7bb87cd05536810cc9fc29d3fbe22cfd359ba190" :: Text))
-          ]
-      in assertEqual "should get encoded message" expectedObject o
 
 -- -------------------------------------------------------------------------- --
 -- tests
